Add files via upload
feat: Added functionality to save and retrieve images to/from the Photo Core Data entity

addPins() was updated to download the Flickr images and save to the Photo entity.  A separate queue is used when downloading the images.  Constraints had to be modified to display the Map and Collection views.